Wall & Werror

+++ src/movelist.cc
@@ -9,14 +9,14 @@ MoveList::MoveList()
 
 void MoveList::PrintMoveList()
 {
-	for(int i = 0; i < this->moves.size(); ++i)
+	for(uint32_t i = 0; i < this->moves.size(); ++i)
 	{
 		Move curMove = this->moves[i];
 		printf("Move: %02d > %s (score: %d) : ",i ,curMove.ToString().c_str(), curMove.score );
 		std::bitset<32> bits(curMove.move);
 		std::cout <<  bits <<std::endl;
 	}
-		printf("MoveList Total: %d Moves\n\n", this->moves.size() );
+		printf("MoveList Total: %lu Moves\n\n", this->moves.size() );
 }
 
 void MoveList::AddQuietMove(Board& pos, Move move)
@@ -32,7 +32,7 @@ void MoveList::AddEnPasMove(Board& pos, Move move)
 	this->moves.push_back(move);
 }
 
-void MoveList::AddPawnMove(Board& pos, int from, int to, int side)
+void MoveList::AddPawnMove(Board& pos, uint32_t from, uint32_t to, uint32_t side)
 {
 	if(side == WHITE)
 	{
@@ -64,7 +64,7 @@ void MoveList::AddPawnMove(Board& pos, int from, int to, int side)
 	}
 
 }
-void MoveList::AddPawnCaptureMove(Board& pos, int from, int to, int cap, int side)
+void MoveList::AddPawnCaptureMove(Board& pos, uint32_t from, uint32_t to, uint32_t cap, uint32_t side)
 {
 	if(side == WHITE)
 	{
@@ -97,13 +97,13 @@ void MoveList::AddPawnCaptureMove(Board& pos, int from, int to, int cap, int sid
 
 }
 
-void MoveList::GenerateBishopMoves(Board& pos, int side)
+void MoveList::GenerateBishopMoves(Board& pos, uint32_t side)
 {
 	int bi = side == WHITE ? wB : bB;
-	for(int pce = 0; pce < pos.piece_list[bi].size(); ++pce )
+	for(uint32_t pce = 0; pce < pos.piece_list[bi].size(); ++pce )
 	{
 		int curBiSq = pos.piece_list[bi][pce];
-		for(int move = 0; move < Attack::BiMoves.size(); ++move )
+		for(uint32_t move = 0; move < Attack::BiMoves.size(); ++move )
 		{
 			int newSq = curBiSq + Attack::BiMoves[move];
 			while(pos.SqOnBoard(newSq))
@@ -123,13 +123,13 @@ void MoveList::GenerateBishopMoves(Board& pos, int side)
 	}
 }
 
-void MoveList::GenerateRookMoves(Board& pos, int side)
+void MoveList::GenerateRookMoves(Board& pos, uint32_t side)
 {
 	int rk = side == WHITE ? wR : bR;
-	for(int pce = 0; pce < pos.piece_list[rk].size(); ++pce)
+	for(uint32_t pce = 0; pce < pos.piece_list[rk].size(); ++pce)
 	{
 		int curRkSq = pos.piece_list[rk][pce];
-		for(int move = 0; move < Attack::RkMoves.size(); ++move)
+		for(uint32_t move = 0; move < Attack::RkMoves.size(); ++move)
 		{
 			int newSq = curRkSq + Attack::RkMoves[move];
 			while(pos.SqOnBoard(newSq))
@@ -148,14 +148,14 @@ void MoveList::GenerateRookMoves(Board& pos, int side)
 		}
 	}
 }
-void MoveList::GenerateQueenMoves(Board& pos, int side)
+void MoveList::GenerateQueenMoves(Board& pos, uint32_t side)
 {
 	int Qn = side == WHITE ? wQ : bQ;
 
-	for(int pce = 0; pce < pos.piece_list[Qn].size(); ++pce)
+	for(uint32_t pce = 0; pce < pos.piece_list[Qn].size(); ++pce)
 	{
 		int curQn = pos.piece_list[Qn][pce];
-		for(int move = 0; move < Attack::BiMoves.size(); ++move)
+		for(uint32_t move = 0; move < Attack::BiMoves.size(); ++move)
 		{
 			int newSq = curQn + Attack::BiMoves[move];
 			while(pos.SqOnBoard(newSq))
@@ -174,10 +174,10 @@ void MoveList::GenerateQueenMoves(Board& pos, int side)
 		}
 	}
 
-	for(int pce = 0; pce < pos.piece_list[Qn].size(); ++pce)
+	for(uint32_t pce = 0; pce < pos.piece_list[Qn].size(); ++pce)
 	{
 		int curQn = pos.piece_list[Qn][pce];
-		for(int move = 0; move < Attack::RkMoves.size(); ++move)
+		for(uint32_t move = 0; move < Attack::RkMoves.size(); ++move)
 		{
 			int newSq = curQn + Attack::RkMoves[move];
 			while(pos.SqOnBoard(newSq))
@@ -196,7 +196,7 @@ void MoveList::GenerateQueenMoves(Board& pos, int side)
 		}
 	}
 }
-void MoveList::GenerateSlidingMoves(Board& pos, int side)
+void MoveList::GenerateSlidingMoves(Board& pos, uint32_t side)
 {
 	this->GenerateBishopMoves(pos, side);
 	this->GenerateRookMoves(pos, side);
@@ -204,13 +204,13 @@ void MoveList::GenerateSlidingMoves(Board& pos, int side)
 }
 
 
-void MoveList::GenerateKnightMoves(Board& pos, int side)
+void MoveList::GenerateKnightMoves(Board& pos, uint32_t side)
 {
 	int kn = side == WHITE ? wN : bN;
-	for(int pce = 0; pce < pos.piece_list[kn].size(); ++pce)
+	for(uint32_t pce = 0; pce < pos.piece_list[kn].size(); ++pce)
 	{
 		int curKnSq = pos.piece_list[kn][pce];
-		for(int move = 0; move < Attack::KnMoves.size(); ++move)
+		for(uint32_t move = 0; move < Attack::KnMoves.size(); ++move)
 		{
 			int newSq = curKnSq + Attack::KnMoves[move];
 			if(pos.SqOnBoard(newSq))
@@ -229,12 +229,12 @@ void MoveList::GenerateKnightMoves(Board& pos, int side)
 }
 
 //*** does not include attacked squares.
-void MoveList::GenerateKingMoves(Board& pos, int side)
+void MoveList::GenerateKingMoves(Board& pos, uint32_t side)
 {
 	int ki = side == WHITE ? wK : bK;
 	int kingSq = pos.piece_list[ki][0];
 	ASSERT(pos.piece_list[ki].size() == 1);
-	for(int move = 0; move < Attack::KiMoves.size(); ++move)
+	for(uint32_t move = 0; move < Attack::KiMoves.size(); ++move)
 	{
 		int newSq = kingSq + Attack::KiMoves[move];
 		if(pos.SqOnBoard(newSq) /*&& !pos.SqAttacked(newSq, !side)*/)
@@ -250,24 +250,24 @@ void MoveList::GenerateKingMoves(Board& pos, int side)
 		}
 	}
 }
-void MoveList::GenerateNonSlidingMoves(Board& pos, int side)
+void MoveList::GenerateNonSlidingMoves(Board& pos, uint32_t side)
 {
 	this->GenerateKnightMoves(pos, side);
 	this->GenerateKingMoves(pos, side);
 }
 
 
-void MoveList::GeneratePawnMoves(Board& pos, int side)
+void MoveList::GeneratePawnMoves(Board& pos, uint32_t side)
 {
-	int oppositeSide = !side;
+	uint32_t oppositeSide = !side;
 
-	int oneMove = side == WHITE ? 10 : -10;
-	int twoMove = side == WHITE ? 20 : -20;
-	int cap1 = side == WHITE ? 9 : -9;
-	int cap2 = side == WHITE ? 11 : -11;
+	int32_t oneMove = side == WHITE ? 10 : -10;
+	int32_t twoMove = side == WHITE ? 20 : -20;
+	int32_t cap1 = side == WHITE ? 9 : -9;
+	int32_t cap2 = side == WHITE ? 11 : -11;
 	int pawn = side == WHITE ? wP : bP;
 
-	for(int pceNum = 0; pceNum < pos.piece_list[pawn].size(); ++pceNum)
+	for(uint32_t pceNum = 0; pceNum < pos.piece_list[pawn].size(); ++pceNum)
 	{
 		int sq = pos.piece_list[pawn][pceNum];
 		ASSERT(sq != OFFBOARD && sq != NO_SQ)
@@ -287,18 +287,18 @@ void MoveList::GeneratePawnMoves(Board& pos, int side)
 		{
 			this->AddPawnCaptureMove(pos, sq, sq + cap2, pos.pieces[sq + cap2], side);
 		}
-		if(sq + cap1 == pos.en_pas)
+		if(sq + cap1 == (int32_t)pos.en_pas)
 		{
 			this->AddCaptureMove(pos, Move(sq, sq + cap1, EMPTY, EMPTY, EP));
 		}
-		if(sq + cap2 == pos.en_pas)
+		if(sq + cap2 == (int32_t) pos.en_pas)
 		{
 			this->AddCaptureMove(pos, Move(sq, sq + cap2, EMPTY, EMPTY, EP));
 		}
 	}
 }
 
-void MoveList::GenerateCastlingMoves(Board& pos, int side)
+void MoveList::GenerateCastlingMoves(Board& pos, uint32_t side)
 {
 	if(side == WHITE)
 	{
